bookglebookgle-ai/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
chroma-hnswlib==0.7.3
chromadb>=0.4.17,<0.7.0
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.5
dataclasses-json==0.6.7
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
durationpy==0.10
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.3
grpcio==1.64.0
grpcio-status==1.64.0
grpcio-tools==1.64.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
img2pdf==0.6.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==33.1.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langsmith==0.4.8
loguru==0.7.3
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
numpy>=1.24.4,<2.0
oauthlib==3.3.1
ocrmypdf==16.10.4
onnxruntime==1.22.1
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.0
overrides==7.7.0
packaging==25.0
pandas>=2.1.4,<2.4
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20250506
pikepdf==9.10.2
pillow>=10.0.1
pluggy==1.6.0
posthog==6.3.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.26.1
pulsar-client==3.8.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.2.3
pypdf==5.8.0
PyPika==0.48.9
pyreadline3==3.5.4;platform_system=="Windows"
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn>=1.3.2,<1.8
scipy>=1.15.0,<1.17
sentence-transformers==5.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sympy==1.14.0
tenacity==9.1.2
tesseract==0.1.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch>=2.2.0,<2.8
tqdm==4.67.1
transformers>=4.45.0,<4.55
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions>=4.8.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
win32_setctime==1.2.0;platform_system=="Windows"
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
psutil==5.9.8
opencv-python-headless>=4.8.0
shapely>=2.0.0
redis>=4.2